% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii_to_tif.R
\name{ascii_to_tif}
\alias{ascii_to_tif}
\title{ASC To Tif Converter}
\usage{
ascii_to_tif(file_full_names)
}
\arguments{
\item{file_full_names}{A vector of file names that are currently ascii to be converted into geotif files.}
}
\description{
Basic converter to turn ascii files into tifs for compressed use within Burn-P3
}
\examples{

\dontrun{

ascii_to_tif(list.files(dir, pattern = ".asc$", full.names = T))

}

}
