% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BP3_Directory_Generator.R
\name{bp3_dir_gen}
\alias{bp3_dir_gen}
\title{Burn-P3 Directory Generator}
\usage{
bp3_dir_gen(root = "", project_name = "")
}
\arguments{
\item{root}{The root directory where the full Burn-P3 directory will be generated.}

\item{project_name}{The name of the project to generate a Burn-P3 project for.}
}
\value{
list
}
\description{
A common directory for Burn-P3 Projects
}
\examples{

dir_list <- bp3_dir_gen(project_name = "Test_Project")
print(dir_list)

}
