% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BP3_Directory.R
\name{bp3_dirs}
\alias{bp3_dirs}
\title{Burn-P3 Directory}
\usage{
bp3_dirs(root = "", project_name = "")
}
\arguments{
\item{root}{The root directory where the full Burn-P3 directory currently exists.}

\item{project_name}{The name of the current Burn-P3 project.}
}
\description{
A function to generate objects with base and Burn-P3 directories for use in subsequent functions.
}
\examples{

bp3_dirs(project_name = "Test_Project")

}
