% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burn_hours.R
\name{burn_hrs}
\alias{burn_hrs}
\title{Hours of Burning Function}
\usage{
burn_hrs(reference_grid, season_df, out_dir)
}
\arguments{
\item{reference_grid}{A reference raster from the area of interest. This will provide the midpoint of the landscape for use when calculating the day length values}

\item{season_df}{A data.frame with 3 columns: season, jstart, jend}

\item{out_dir}{Output directory}
}
\value{

}
\description{
The hours of burning are an important part of a Burn-P3 simulation. One third of the total daylight hours are considered to be burning hours. Values are rounded to the nearest whole number and the proportion of each of those hours will be output to a file in your declared location. If you are declaring different hours for the whole simulation, use a single season with the start and end julian days spanning the weather relevant to your study.
}
\examples{

elev <- rast(system.file("extdata","elev.tif",package="BurnP3.HelpR"))

season_df <- data.frame(season = c("Spring","Summer"),
                        jstart = c(125,160),
                        jend = c(159, 185))

out_dir <- tempdir()

burn_hrs( reference_grid = elev,
          season_df = season_df,
          out_dir = out_dir)

print(paste0("Files can be found at: ",gsub("\\\\\\\\","/",out_dir)))

}
