% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BurnP3-package.R
\docType{package}
\name{BurnP3-package}
\alias{BurnP3}
\alias{BurnP3-package}
\title{BurnP3: Burn-P3 Input Generation Package}
\description{
A series of functions intended to aid experience Burn-P3 operators in the generation of inputs for a modelling task.
}
\keyword{internal}
