% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_ign.R
\name{density_ign}
\alias{density_ign}
\title{Kernal Density Calculator}
\usage{
density_ign(reference_grid, layer, output_location, output_name)
}
\arguments{
\item{reference_grid}{This is a reference raster to provide a projection and a surface to assign values onto, this should be a grid that registers with the other grids you are using for your project.}

\item{layer}{A point layer for which you are interested in the density.}

\item{output_location}{The location the output density rasters will be placed.}

\item{output_name}{The name of the files to be output. Typically these will be a vector of the names of your list of layers.}
}
\value{
RasterLayer
}
\description{
Generate kernal density layers from a list of layers and write the rasters out
}
\details{
This function utilizes the kernal density estimation algorithm from spatialEco to create raster layers for use in ignition grid calculation.
}
\seealso{
\code{\link[BurnP3]{spatdat_ign_layer}}

\code{\link[spatialEco]{sp.kde}}
}
\author{
Brett Moore, \email{Brett.Moore@canada.ca}
}
