% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_ign.R
\name{distance_ign}
\alias{distance_ign}
\title{Euclidean Distance Calculator}
\usage{
distance_ign(reference_grid, rasters_list, output_location, output_name)
}
\arguments{
\item{reference_grid}{This is a reference raster to provide a projection and a surface to assign values onto, this should be a grid that registers with the other grids you are using for your project.}

\item{rasters_list}{The location the output density rasters will be placed.}

\item{output_location}{The name of the files to be output. Typically these will be a vector of the names of your list of layers.}

\item{output_name}{}
}
\value{
RasterLayer
}
\description{
Generate kernal density layers from a list of layers and write the rasters out
}
\details{
This function generates a euclidean distance grid with the \code{distance} function.
}
\author{
Brett Moore, \email{Brett.Moore@canada.ca}
}
