% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elev_grab.R
\name{elev_grab}
\alias{elev_grab}
\title{Elevation Grabber - From NTS Grids}
\usage{
elev_grab(reference_grid, output_directory)
}
\arguments{
\item{reference_grid}{This is a reference raster to provide a projection and a surface to assign values onto, this should be a grid that registers with the other grids you are using for your project.}

\item{output_directory}{Output directory where your elevation and Wind Ninja elevation grids will be stored.}
}
\value{
RasterLayer
}
\description{
This function uses your reference layer to define the NTS grids necessary to pull elevation from the CDEM layers provided by the Government of Canada.
}
\details{
The purpose of this function is to generate a common and rapid elevation layer that is sampled and masked to the reference grid for use within Burn-P3. A second elevation grid is also generated for use in Wind Ninja as that software will fail with NA values in the elevation grid.
}
\references{
\href{https://open.canada.ca/data/en/dataset/7f245e4d-76c2-4caa-951a-45d1d2051333}{Open Data Canada}

[Open Data Canada FTP]{http://ftp.maps.canada.ca/pub/nrcan_rncan/vector/index/}

[Wind Ninja]{https://www.firelab.org/project/windninja}
}
