% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r, R/fire_data.r
\docType{data}
\name{fire_data}
\alias{fire_data}
\title{fire_data}
\format{
a \code{sf::SpatialPointsDataframe} object with typical fire
information used when generating ignition grids with 120 rows and 30 columns

\describe{
\item{"SRC_AGENCY"}{Source Agency}
\item{"FIRE_ID"}{Agency Fire Number}
\item{"FIRENAME"}{Agency Fire Name (if available)}
\item{"LATITUDE"}{Latitude}
\item{"LONGITUDE"}{Longitude}
\item{"YEAR"}{Year}
\item{"MONTH"}{Month}
\item{"DAY"}{Day}
\item{"REP_DATE"}{Date fire was reported}
\item{"ATTK_DATE"}{Date fire was initially actioned}
\item{"OUT_DATE"}{Date fire was declared out}
\item{"DECADE"}{Decade the fire occurred in}
\item{"SIZE_HA"}{Area burned as reported by the agency in hectares}
\item{"CAUSE"}{Agency declared fire cause}
\item{"PROTZONE"}{Type of protection zone (full response / modified / etc.)}
\item{"FIRE_TYPE"}{Flaming behaviour category, surface, intermittent crown, crown}
\item{"MORE_INFO"}{Additional Information}
\item{"CFS_REF_ID"}{Federal Fire Number}
\item{"CFS_NOTE1"}{Note}
\item{"CFS_NOTE2"}{Note}
\item{"ACQ_DATE"}{Acquisition date}
\item{"ECODISTRIC"}{Ecological District}
\item{"ECOREGION"}{Ecological Region}
\item{"ECOZONE"}{Ecological Zone}
\item{"CFS_ECOZ"}{Ecological Zone - Canadian Forest Service}
\item{"jday"}{Julian Day}
\item{"season"}{Season (1-3 for spring, summer, and fall.)}
}

\subsection{\code{fire_data}}{

A data frame with 120 rows and 30 columns
\describe{
\item{"SRC_AGENCY"}{Source Agency}
\item{"FIRE_ID"}{Agency Fire Number}
\item{"FIRENAME"}{Agency Fire Name (if available)}
\item{"LATITUDE"}{Latitude}
\item{"LONGITUDE"}{Longitude}
\item{"YEAR"}{Year}
\item{"MONTH"}{Month}
\item{"DAY"}{Day}
\item{"REP_DATE"}{Date fire was reported}
\item{"ATTK_DATE"}{Date fire was initially actioned}
\item{"OUT_DATE"}{Date fire was declared out}
\item{"DECADE"}{Decade the fire occurred in}
\item{"SIZE_HA"}{Area burned as reported by the agency in hectares}
\item{"CAUSE"}{Agency declared fire cause}
\item{"PROTZONE"}{Type of protection zone (full response / modified / etc.)}
\item{"FIRE_TYPE"}{Flaming behaviour category, surface, intermittent crown, crown}
\item{"MORE_INFO"}{Additional Information}
\item{"CFS_REF_ID"}{Federal Fire Number}
\item{"CFS_NOTE1"}{Note}
\item{"CFS_NOTE2"}{Note}
\item{"ACQ_DATE"}{Acquisition date}
\item{"ECODISTRIC"}{Ecological District}
\item{"ECOREGION"}{Ecological Region}
\item{"ECOZONE"}{Ecological Zone}
\item{"CFS_ECOZ"}{Ecological Zone - Canadian Forest Service}
\item{"jday"}{Julian Day}
\item{"season"}{Season (1-3 for spring, summer, and fall.)}
}
}
}
\usage{
fire_data

fire_data
}
\description{
Examples datasets for the Burn P3 Helpr Package

Examples datasets for the Burn P3 Helpr Package
}
\keyword{datasets}
