% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_rate_distribution.R
\name{fire_rate_distribution}
\alias{fire_rate_distribution}
\title{Fire Rate Distribution}
\usage{
fire_rate_distribution(
  input,
  date_col,
  date_format = "\%Y/\%m/\%d",
  aoi,
  output_location,
  seasonal = F,
  zonal = F,
  seasons = "",
  zones,
  zone_names = "",
  min_fire_size = 0.01,
  causes = c("H", "L")
)
}
\arguments{
\item{input}{Location of a binary file containing fire data, requires a date column}

\item{date_col}{Character string defining the name of the column containing date information.}

\item{date_format}{Format of the dates within the date column that will enable the function to transition the date to a julian day. \emph{(Default = "\%Y/\%m/\%d")}}

\item{aoi}{A spatial object do be used for its CRS and extent. Data will be cropped to this aoi.}

\item{output_location}{Folder location for fire rate distribution to be exported to. NOTE: Only requires the base directory, assumes the directory generator was used.}

\item{seasonal}{Declaration of the use of seasons in the weather data set. \emph{(Default = F)}}

\item{zonal}{Declaration of the use of weather zones. \emph{(Default = F)}}

\item{seasons}{If seasonal is True, a two column data.frame that contains the seasons numerical identifier and the description. \emph{(Default = "")}}

\item{zones}{A raster containing the fire zones to be used.}

\item{zone_names}{If zonal is True, identify the descriptive names of the zones for use during mapping and output. \emph{(Default = "")}}

\item{min_fire_size}{A minimum fire size to subset the fire information to for adjuste fire rate distribution depending on the question being asked. \emph{(Default = 0.01)}}

\item{causes}{A character vector defining the causes within the fire dataset. \emph{(Default = c("H","L"))}}
}
\value{

}
\description{
Fire Rate Distribution
}
\examples{
## Load relavent data
data("fire_data")
aoi <- readOGR(system.file("extdata","extdata.gpkg",package="BurnP3"),"aoi")
output_location <- paste0(tempdir(),"\\\\")
zones <- raster(system.file("extdata","zones.tif",package="BurnP3"))
data("seasons")
zone_names = c("Alpine-E","Montane-E","Alpine-W","Montane-W","IDF")

fr <- fire_rate_distribution(input = fire_data,
                             date_format = "\%Y/\%m/\%d",
                             aoi = aoi,
                             output_location = output_location,
                             date_col = "REP_DATE",
                             seasonal = F,
                             zonal = F,
                             seasons = "",
                             zones,
                             zone_names = "",
                             min_fire_size = 0.01,
                             causes = c("H","L")
                             )
print(fr)

fr <- fire_rate_distribution(input = fire_data,
                             date_format = "\%Y/\%m/\%d",
                             aoi = aoi,
                             output_location = output_location,
                             date_col = "REP_DATE",
                             seasonal = T,
                             zonal = F,
                             seasons = seasons,
                             zones,
                             zone_names = "",
                             min_fire_size = 0.01,
                             causes = c("H","L")
                             )
print(fr)

fr <- fire_rate_distribution(input = fire_data,
                             date_format = "\%Y/\%m/\%d",
                             aoi = aoi,
                             output_location = output_location,
                             date_col = "REP_DATE",
                             seasonal = F,
                             zonal = T,
                             seasons = "",
                             zones = zones,
                             zone_names = zone_names,
                             min_fire_size = 0.01,
                             causes = c("H","L")
                             )

print(fr)

unlink(output_location, recursive = T)

}
\seealso{
\link {bp3_dir_gen}\link{BurnP3}
}
