% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_grab.R
\name{grid_grab}
\alias{grid_grab}
\title{Gridded Data Grabber - From NTS Grids}
\usage{
grid_grab(
  aoi_e = NULL,
  buffer = NULL,
  reference_grid = NULL,
  output_directory,
  fuel = TRUE,
  ref_is_fuel = FALSE
)
}
\arguments{
\item{aoi_e}{Polygon or numeric coordinate pair ("longitude","latitude") for the area of interest to be intersected with the NTS grid layer. \emph{(Default: NULL)} If Default is used the full extent under the reference_grid will be returned.}

\item{buffer}{A buffer in meters to define an area of interest around a point if coordinates are given to the \code{aoi_e} parameter. \emph{(Default: NULL)}}

\item{reference_grid}{This is a reference raster to provide a projection and a surface to assign values onto, this should be a grid that registers with the other grids you are using for your project. Can be either the location of the raster or a raster object.}

\item{output_directory}{Output directory where your elevation and Wind Ninja elevation grids will be stored.}

\item{fuel}{If you want the tool to process a fuel grid, set this to TRUE. \emph{(Default: TRUE)}}

\item{ref_is_fuel}{If the reference grid is the fuel grid se this to TRUE. \emph{(Default: FALSE)}}
}
\value{
SpatRast
}
\description{
This function uses your reference layer to define the NTS grids necessary to pull elevation from the CDEM layers and the national FBP Fuel grid provided by the Government of Canada.
}
\details{
The purpose of this function is to generate a common and rapid elevation layer that is sampled and masked to the reference grid for use within Burn-P3. A second elevation grid is also generated for use in Wind Ninja as that software will fail with NA values in the elevation grid. Fuel data from:
}
\note{
Requires an internet connection
}
\examples{
## Load example data
ref_grid <- terra::rast(system.file("extdata/fuel.tif",package = "BurnP3.HelpR"))
temp_dir <- tempdir()
output_directory <- paste0(temp_dir,"\\\\")
test <- grid_grab(reference_grid = ref_grid,
                  output_directory = output_directory)

## use with a coordinate pair
test2 <- grid_grab(aoi_e = c(-110, 54),
                  buffer = 50000,
                  reference_grid = ref_grid,
                  output_directory = output_directory)

## use with a coordinate pair but no buffer
test3 <- grid_grab(aoi_e = c(-110, 54),
                  reference_grid = ref_grid,
                  output_directory = output_directory)

## use with a coordinate pair
test4 <- grid_grab(reference_grid = ref_grid,
                  output_directory = output_directory,
                  ref_is_fuel=T)

## If the reference grid is outside the location desired, it will return NAN
## use with a coordinate pair
test5 <- grid_grab(aoi_e = c(-110, 54),
                  buffer = 50000,
                  reference_grid = ref_grid,
                  output_directory = output_directory,
                  ref_is_fuel=T)

unlink(temp_dir)
}
\references{
\href{https://open.canada.ca/data/en/dataset/4e66dd2f-5cd0-42fd-b82c-a430044b31de/resource/f645a02d-fbb2-4c35-9bec-8d4ff516f5f8}{National FBP Fuels}

\href{https://open.canada.ca/data/en/dataset/7f245e4d-76c2-4caa-951a-45d1d2051333}{Open Data Canada}

\href{http://ftp.maps.canada.ca/pub/nrcan_rncan/vector/index/}{Open Data Canada FTP}

\href{https://www.firelab.org/project/windninja}{Wind Ninja}
}
