% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_mapping.R
\name{haz_mapping}
\alias{haz_mapping}
\title{Hazard Mapping Utility}
\usage{
haz_mapping(fi, bp, bp_break_divisor = 4, hfi_breaks = c(0, 2000, 4000, 10000))
}
\arguments{
\item{fi}{Fire intensity spatrast}

\item{bp}{Burn Probability spatrast (must be the burn probability and not the burn count)}

\item{bp_break_divisor}{A divisor for the burn probability break, this will dictate the number of HFI breaks required. \emph{Default: 4}}

\item{hfi_breaks}{A character vector of hfi breaks for use within the mapping system.}
}
\value{
SpatRast
}
\description{
This utility will use a fire intensity grid and a burn probability grid and output a hazard grid.
}
