% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_legend.R
\name{hazard_legend}
\alias{hazard_legend}
\title{Hazard Legend}
\usage{
hazard_legend(
  output_location,
  output_filename,
  hfi_breaks = c(0, 2000, 4000, 10000),
  prob_vals = seq(25, 100, 25)
)
}
\arguments{
\item{output_location}{Output location}

\item{output_filename}{Output filename with png extension}

\item{hfi_breaks}{A string of 4 numeric HFI breaks}
}
\value{

}
\description{
This function generates a 4x4 legend square for the quartiles of probability and a set of 4 HFI breaks.
}
