% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_rast.R
\name{hazard_rast}
\alias{hazard_rast}
\title{Wildfire Hazard Raster Generator}
\usage{
hazard_rast(
  stats_file,
  median_pattern,
  mean_pattern,
  burn_count_pattern,
  intensity_directory,
  burn_count_directory
)
}
\arguments{
\item{stats_file}{Character string to the location of the statistics file, if there have been multiple runs aggregated, the aggregate statistics file.}

\item{burn_count_directory}{}
}
\value{

}
\description{
Wildfire Hazard Raster Generator
}
