% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ign_grid.R
\name{ign_grid}
\alias{ign_grid}
\title{Ignition Grids}
\usage{
ign_grid(
  fire_data,
  indicator_stack,
  indicators_1,
  indicators_2,
  causes,
  season_description,
  output_location,
  min_fire_size = "",
  model = ""
)
}
\arguments{
\item{fire_data}{Spatial Points Data Frame containing wildfire location data, this should be cropped to the area of interest or it will take a very long time to write NA values into the data.}

\item{indicator_stack}{Raster Stack of explanatory variables within the area of interest.}

\item{indicators_1}{Character vector describing the layer name from the indicator stack that should be used when calculating ignition probability for the first cause.}

\item{indicators_2}{Character vector describing the layer name from the indicator stack that should be used when calculating ignition probability for the second cause.}

\item{causes}{Character vector describing the causes within the data, typically H (human) and L (lightning), used for naming files and filtering wildfire data.}

\item{season_description}{Character vector with the descriptive names of the factor seasons within the wildfire data.}

\item{output_location}{Directory for the rasters that will be calculated.}

\item{min_fire_size}{A minimum fire size that may be defined to filter the wildfire data provided. \emph{(Default = "")}}

\item{model}{A character string defining the model to be used during the ignition gridding process. Can be one of: rf_stock, rf, gbm. The models are: stock random forest - run without any tuning. Random forest, an automatically tuned random forest run. Gradient Boosted Model - a gbm that is run in its default mode.}
}
\value{

}
\description{
Ignition Grids
}
