% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{indicator_stack}
\alias{indicator_stack}
\title{Indicator stack}
\format{
A \code{terra::SpatRaster} with 21 layers.
Resolution: 100 x 100; Extent: xmin=440200.0000, xmax=465700.0000,
ymin=5700500.0000, ymax=5734000.0000;
CRS: +proj=tmerc +lat_0=0 +lon_0=-115 +k=0.9992 +x_0=500000 +y_0=0
+ellps=GRS80 +units=m +no_defs

\describe{

\item{"elevation"}{Elevation (m)}
\item{"fuels"}{FBP Fuel Type}
\item{"ecodistrict"}{Ecological District}
\item{"topography_position_index"}{Topographic Position Index (TPI)}
\item{"terrain_ruggedness_index"}{Terrain Ruggedness Index (TRI)}
\item{"road_density"}{Road Density}
\item{"trail_density"}{Trail Density}
\item{"asset_distance"}{Euclidian Distance to Assets}
\item{"trail_distance"}{Euclidian Distance to Trails}
\item{"road_distance"}{Euclidian Distance to Roads}
\item{"solar_rad"}{Incident Solar Radiation}
\item{"lightning_density"}{Density of Lightning Strikes}
\item{"asset_density"}{Asset Density}
\item{"rail_distance"}{Distance to Railway}
\item{"rail_density"}{Density of Railway}
\item{"in_out_park"}{In or Out of a National Park}
\item{"river_distance"}{Distance to River}
\item{"river_density"}{Density of River}
\item{"lat"}{Latitude}
\item{"lon"}{Longitude}
\item{"town_boundary"}{In or Out of the boundary of a town of interest}

}
}
\usage{
indicator_stack
}
\description{
Example raster data (terra) of a set of indicators for building ignition grids.
}
\examples{
data(indicator_stack)
indicator_stack <- terra::unwrap(indicator_stack)
if (requireNamespace("terra", quietly = TRUE)) {
  terra::nlyr(indicator_stack); terra::names(indicator_stack)
}

}
\keyword{datasets}
