% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{interest_raster}
\alias{interest_raster}
\title{Interest raster}
\format{
A \code{terra::SpatRaster} with 1 layer.
Layer names: fuel
Resolution: 99.9999936090549 x 99.9999936090754; Extent: xmin=440201.0086,
xmax=465701.0070, ymin=5700499.2515, ymax=5733999.2493;
CRS: +proj=tmerc +lat_0=0 +lon_0=-115 +k=0.9992 +x_0=500000 +y_0=0
+ellps=GRS80 +units=m +no_defs
}
\usage{
interest_raster
}
\description{
Example raster data (terra) of an area of interest.
}
\examples{
data(interest_raster)
interest_raster <- terra::unwrap(interest_raster)
if (requireNamespace("terra", quietly = TRUE)) {
  terra::nlyr(interest_raster); terra::names(interest_raster)
}

}
\keyword{datasets}
