% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{lut}
\alias{lut}
\title{Lut}
\format{
A data frame with 154 rows and 10 columns:
\describe{
\item{grid_value}{Value in the grid}
\item{export_value}{Value the grid should be interpretted as by Promtheus}
\item{descriptive_name}{Descriptive FBP Fuel Type}
\item{fuel_type}{FBP Fuel Type}
\item{r}{red}
\item{g}{green}
\item{b}{blue}
\item{h}{hue}
\item{s}{saturation}
\item{l}{luminance}
}
}
\usage{
lut
}
\description{
Example look up table for FBP Fuel Types in Prometheus data.
}
\examples{
data(lut)
head(lut)

}
\keyword{datasets}
