% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_bp_classification.R
\name{mean_bp_classification}
\alias{mean_bp_classification}
\title{Mean Burn Probability Classification}
\usage{
mean_bp_classification(input, output_filename)
}
\arguments{
\item{input}{Required as a SpatRast or a character string for the directory of the raster information.}

\item{output_filename}{A character sting that defines the output directory and filename for the output raster.}
}
\value{
SpatRaster
}
\description{
This function creates an output raster that represents the deviation from the mean burn probability on a landscape. This is intended for use when presenting burn probability to better demonstrate the deviance from the mean across the landscape and draw attention to areas that are considerably above and below the mean burn probability.
}
