% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrdem_helper.R
\name{mrdem_subset_stac}
\alias{mrdem_subset_stac}
\title{\subsection{MRDEM Subset Stac}{

MRDEM-30 subset & derivatives from the Government of Canada Data Cube.
Robust to occasionally-missing .tif assets (auto-fallback to .vrt & local compute).
Subset MRDEM-30 DTM by bbox or centroid in any CRS; optionally fetch slope/aspect/hillshade
from STAC or compute locally; write COGs w/ overviews; optional aligned & exact outputs.
}}
\usage{
mrdem_subset_stac(
  output_base,
  bbox = NULL,
  centroid = NULL,
  aoi_crs = NULL,
  buffer_m = NULL,
  return_versions = c("exact_aeqd", "aligned_source"),
  derivatives = c("slope", "aspect"),
  prefer_remote_derivatives = FALSE,
  dtm_href = NULL,
  stac_url = "https://datacube.services.geo.ca/stac/api/",
  target_res_m = 100,
  resampling = "bilinear",
  write_cog = TRUE,
  cog_options = list(gdal = c("COMPRESS=LZW", "BLOCKSIZE=512", "OVERVIEWS=AUTO",
    "RESAMPLING=BILINEAR", "NUM_THREADS=ALL_CPUS", "BIGTIFF=IF_SAFER")),
  overwrite = TRUE,
  ignore_ssl = FALSE
)
}
\arguments{
\item{bbox}{Bounding box for the area of interest. \emph{Default=NULL}}

\item{centroid}{If no bounding box is provided, a centroid can be used. \emph{Default=NULL}}

\item{aoi_crs}{The coordinate reference system (Projection) of the area of interest. \emph{Default=NULL}}

\item{buffer_m}{Distance in meters to buffer the centroid by. \emph{Default=NULL}}

\item{dtm_href}{The tag for digital terrain model. \emph{Default=NULL}}

\item{stac_url}{The location of the data. \emph{Default="https://datacube.services.geo.ca/stac/api/"}}

\item{target_res_m}{The target resolution in meters. \emph{Default=100}}

\item{ignore_ssl}{A flag to allow a download without ssl (this should only be used for testing, fix your network, don't step around it.) \emph{Default=FALSE}}

\item{target_crs}{The target coordinate reference system. (Projection) \emph{Default=NULL}}

\item{use_vsimem}{Do the work in memory. \emph{Default=FALSE}}
}
\value{
SpatRast
}
\description{
\subsection{MRDEM Subset Stac}{

MRDEM-30 subset & derivatives from the Government of Canada Data Cube.
Robust to occasionally-missing .tif assets (auto-fallback to .vrt & local compute).
Subset MRDEM-30 DTM by bbox or centroid in any CRS; optionally fetch slope/aspect/hillshade
from STAC or compute locally; write COGs w/ overviews; optional aligned & exact outputs.
}
}
