% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrdem_helper.R
\name{mrdem_subset_windowed}
\alias{mrdem_subset_windowed}
\title{\subsection{MRDEM Windowed}{
}

\subsection{MRDEM-30 subset & derivatives from the Government of Canada Data Cube.
Robust to occasionally-missing .tif assets (auto-fallback to .vrt & local compute).}{
}}
\usage{
mrdem_subset_windowed(
  bbox = NULL,
  centroid = NULL,
  buffer_m = NULL,
  aoi_crs = NULL,
  target_res_m = 100,
  target_crs = NULL,
  stac_url = "https://datacube.services.geo.ca/stac/api/",
  dtm_href = NULL,
  ignore_ssl = FALSE,
  use_vsimem = F
)
}
\arguments{
\item{bbox}{Bounding box for the area of interest. \emph{Default=NULL}}

\item{centroid}{If no bounding box is provided, a centroid can be used. \emph{Default=NULL}}

\item{buffer_m}{Distance in meters to buffer the centroid by. \emph{Default=NULL}}

\item{aoi_crs}{The coordinate reference system (Projection) of the area of interest. \emph{Default=NULL}}

\item{target_res_m}{The target resolution in meters. \emph{Default=100}}

\item{target_crs}{The target coordinate reference system. (Projection) \emph{Default=NULL}}

\item{stac_url}{The location of the data. \emph{Default="https://datacube.services.geo.ca/stac/api/"}}

\item{dtm_href}{The tag for digital terrain model. \emph{Default=NULL}}

\item{ignore_ssl}{A flag to allow a download without ssl (this should only be used for testing, fix your network, don't step around it.) \emph{Default=FALSE}}

\item{use_vsimem}{Do the work in memory. \emph{Default=FALSE}}
}
\value{
SpatRast
}
\description{
\subsection{MRDEM Windowed}{
}

\subsection{MRDEM-30 subset & derivatives from the Government of Canada Data Cube.
Robust to occasionally-missing .tif assets (auto-fallback to .vrt & local compute).}{
}
}
