% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{season_df}
\alias{season_df}
\title{Season df}
\format{
A data frame with 6 rows and 3 columns:
\describe{
\item{jday}{\code{numeric}}
\item{season_desc}{\code{character}}
\item{season}{Levels present: \code{1, 2, 3, 4, 5, 6} Corresponding with
Early / Late Spring, Summer, and Fall}
}
}
\usage{
season_df
}
\description{
Example seasonal tabular data.
}
\examples{
data(season_df)
head(season_df)

}
\keyword{datasets}
