% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatdat_ign_layer.R
\name{spatdat_ign_layer}
\alias{spatdat_ign_layer}
\title{Spatial Data Importer}
\usage{
spatdat_ign_layer(reference_grid, layer, dsn)
}
\arguments{
\item{reference_grid}{This is a reference raster to provide a projection and a surface to assign values onto, this should be a grid that registers with the other grids you are using for your project.}

\item{layer}{Names of the input layers for import, allows the system to function if there are multiple shapes at a single data source name.}

\item{dsn}{Locations of the shapefiles to be loaded in and operated upon. Also accepts a list containing sf objects.}
}
\value{
List
}
\description{
Import spatial data for rasterization in the ignition grid development process.
}
\details{
This tool imports the spatial data for use within the spatial data rasterizer, it can perform on points, lines and polygons.
}
\examples{

## Load example data
ref_grid <- rast(system.file("extdata/fuel.tif",package = "BurnP3.HelpR"))

out <- spatdat_ign_layer(reference_grid = ref_grid,
                         layer = "road",
                         dsn = system.file("extdata/extdata.gpkg",package="BurnP3.HelpR"))

}
