% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatdat_ign_rast.R
\name{spatdat_ign_rast}
\alias{spatdat_ign_rast}
\title{Spatial Data Rasterizer}
\usage{
spatdat_ign_rast(reference_grid, layers_list)
}
\arguments{
\item{reference_grid}{This is a reference raster to provide a projection and a surface to assign values onto, this should be a grid that registers with the other grids you are using for your project.}

\item{layer}{A list containing at least one spatial layer from the \code{spatdat_ign_layer} function to be rasterized.}
}
\value{
List containing points and raster layers.
}
\description{
A function to create a point layer and a raster layer out of spatial inputs (point, line, polygon) for use during ignition grid generation. These layers are intended for use in the density \code{density_ign} and euclidean distance \code{distance_ign} functions.
}
\examples{

## Load example data
ref_grid <- raster(system.file("extdata/fuel.tif",package = "BurnP3"))
layers_list <- list(readOGR(dsn=system.file("extdata/extdata.gpk",package = "BurnP3"),layer="layers_list"))
out <- spatdat_ign_rast(reference_grid = ref_grid,
                        layers_list = layers_list)

}
\seealso{
\code{\link[BurnP3]{spatdat_ign_layer}}
}
