\name{test_rast_hour02}
\docType{data}
\alias{test_rast_hour02}
\title{Raster Data for ffmcRaster function}
\description{
  Hourly fire weather inputs obtained from the Global Environmental Multiscale Model (GEM) in northern Alberta
}
\usage{system.file("extdata", test_rast_hour02, package="cffdrs")}
\format{A raster (tif) file.}
\keyword{datasets}
