% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{weather}
\alias{weather}
\title{Weather}
\format{
A data frame with 3270 rows and 23 columns:
\describe{
\item{date}{\code{POSIXct} Date of the weather record}
\item{temp}{\code{numeric} Temperature(°C)}
\item{rh}{\code{numeric} Relative Humidity(\%)}
\item{ws}{\code{numeric} Wind Speed (km/hr)}
\item{wd}{\code{numeric} Wind Direction (°)}
\item{prec}{\code{numeric} Precipitation (mm)}
\item{ffmc}{\code{numeric} Fine Fuel Moisture Code}
\item{dmc}{\code{numeric} Duff Moisture Code}
\item{dc}{\code{numeric} Drought Code}
\item{bui}{\code{numeric} Build Up Index}
\item{isi}{\code{numeric} Initial Spread Index}
\item{fwi}{\code{numeric} Fire Weather Index}
\item{dsr}{\code{numeric} Daily Severity Rating}
\item{yr}{\code{numeric} Year}
\item{mon}{\code{numeric} Month}
\item{day}{\code{numeric} Day}
\item{id}{\code{factor} Weather Station ID}
\item{station_name}{\code{character} Weather Station Name}
\item{lat}{\code{numeric} Latitude}
\item{long}{\code{numeric} Longitude}
\item{wx_zone}{\code{numeric} Weather Zone}
\item{Class}{\code{factor}}
\item{season}{Levels present: \code{1, 2, 3, 4, 5, 6} Representing Early / Late Spring, Summer, and Fall}
}
}
\usage{
weather
}
\description{
Example weather data table.
}
\examples{
data(weather)
head(weather)

}
\keyword{datasets}
