% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wn_converter.R
\name{wn_converter}
\alias{wn_converter}
\title{Wind Ninja ASCII to TIF Converter}
\usage{
wn_converter(directory)
}
\arguments{
\item{directory}{directory of wind ninja grids. The directory that contains the direction and velocity folders.}
}
\description{
A quick converter for ASCII wind ninja grids to GeoTiff for smaller storage and faster loading.
}
