% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi.R
\name{aoi}
\alias{aoi}
\title{Area of Interest for Weather Stations - Designed for Parks Canada Data}
\usage{
aoi(
  area_of_interest_file,
  reference_grid,
  buffer_width = 0,
  park_of_interest = "",
  stns,
  stn_name_col,
  stn_id_col
)
}
\arguments{
\item{area_of_interest_file}{Two element vector describing the location of the file (dsn) and the filename. The file is expected to follow a format readable by readOGR. - In the future may have automatic Parks Canada download
TODO: [WORK-26] add PC functionality}

\item{reference_grid}{This is a reference raster to provide a projection as well as a graphic representation of your area of interest in order to see the weather stations within the area.}

\item{buffer_width}{A width in meters to apply to the area of interest. If you are only interested in the area do not buffer.}

\item{stns}{Dataframe containing the stations within and around your area of interest.}

\item{stn_name_col}{A character string containing the column of the station names}

\item{stn_id_col}{A character string containing the column of the station ids}

\item{park_of_interest=""}{The english name of the park or parks of interest. Expects a vector of character strings.}
}
\description{
Area of Interest for Weather Stations - Designed for Parks Canada Data
}
\author{
Brett Moore, \email{Brett.Moore@canada.ca}
}
