% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_output.R
\name{combine_output}
\alias{combine_output}
\title{Combine Output}
\usage{
combine_output(
  directory,
  file_prefix = "Combined",
  stats_file,
  bp_file,
  polygon = F,
  raster = T,
  daily = F
)
}
\arguments{
\item{directory}{Directory where the outputs are stored and new outputs will be generated to.}

\item{file_prefix}{A file prefix that will precede "stats", "polygons" and/or "burn probability". \emph{(Default: "Combined")}}

\item{stats_file}{A text string identifying the name of the statistics file. Can be left blank if your statistics file ends with 'Statistics'.}

\item{bp_file}{A text string identifying the name of the burn probability file. Can be left blank if your burn probability file ends with 'Probability'.}

\item{polygon}{A logical flag for combination of fire polygons. \emph{(Default: False)}}

\item{raster}{A logical flag for combination of burn probability rasters. \emph{(Default: True)}}

\item{daily}{A logical flag for the combination of daily or final perimeters. It will default to final when daily = F. \emph{(Default: False)}}
}
\description{
Combine Output
}
