% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_output.R
\name{combine_output}
\alias{combine_output}
\title{Combine Output}
\usage{
combine_output(
  directory,
  file_prefix = "Combined",
  polygon = F,
  raster = T,
  daily = F
)
}
\arguments{
\item{directory}{Directory where the outputs are stored and new outputs will be generated to.}

\item{file_prefix}{A file prefix that will preceed "stats", "polygons" and/or "burn probability". \emph{(Default: "Combined")}}

\item{polygon}{A logical flag for combination of fire polygons. \emph{(Default: False)}}

\item{raster}{A logical flag for combination of burn probability rasters. \emph{(Default: True)}}

\item{daily}{A logical flag for the combination of daily or final perimeters. It will default to final when daily = F. \emph{(Default: False)}}
}
\value{

}
\description{
Combine Output
}
\examples{

}
