% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_ign.R
\name{density_ign}
\alias{density_ign}
\title{Kernal Density Calculator}
\usage{
density_ign(
  reference_grid,
  layers_list,
  width = 1000,
  output_location,
  output_name
)
}
\arguments{
\item{reference_grid}{Reference raster to provide a projection and a surface to assign values onto, this should be a grid that registers with the other grids you are using for your project.Can be either the location of the raster or a raster object.}

\item{width}{Distance around each point to take into the density calculation.}

\item{output_location}{The location the output density rasters will be placed.}

\item{output_name}{The name of the files to be output. Typically these will be a vector of the names of your list of layers.}

\item{layer}{A point layer for which you are interested in the density.}
}
\value{
RasterLayer
}
\description{
Generate kernal density layers from a list of layers and write the rasters out
}
\details{
This function utilizes the kernal density estimation algorithm from spatialEco to create raster layers for use in ignition grid calculation.
}
\examples{

## Load example data
ref_grid <- raster(system.file("extdata/fuel.tif",package = "BurnP3.HelpR"))
fires <- readOGR(system.file("extdata/fires.shp",package = "BurnP3.HelpR"))
temp_dir <- tempdir()
density_test <- density_ign(reference_grid = ref_grid,
                            layers_list = fires,
                            width = 1000,
                            output_location = temp_dir,
                            output_name = "density_test")

plot(density_test)

unlink(temp_dir)
}
\seealso{
\code{\link[BurnP3.HelpR]{spatdat_ign_layer}}
}
