% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_ign.R
\name{distance_ign}
\alias{distance_ign}
\title{Euclidean Distance Calculator}
\usage{
distance_ign(reference_grid, rasters_list, output_location, output_name)
}
\arguments{
\item{reference_grid}{This is a reference spatraster to provide a projection and a surface to assign values onto, this should be a grid that registers with the other grids you are using for your project.Can be either the location of the raster or a raster object.}

\item{rasters_list}{The location the output density rasters will be placed.}

\item{output_location}{The name of the files to be output. Typically these will be a vector of the names of your list of layers.}

\item{output_name}{Name of the output file}
}
\value{
SpatRaster
}
\description{
Generate euclidean distance layers from a list of spatrasters and write the spatrasters out
}
\details{
This function generates a euclidean distance grid with the \code{distance} function.
}
\examples{
## Load example data
ref_grid <- terra::rast(system.file("extdata/fuel.tif",package = "BurnP3.HelpR"))
data("spatdat_rast_out")
raster_list$rasters <- terra::unwrap(raster_list$rasters)
temp_dir <- tempdir()
distance_test <- distance_ign(reference_grid = ref_grid,
                            rasters_list = raster_list$rasters,
                            output_location = temp_dir,
                            output_name = "distance_test")

terra::plot(distance_test)

unlink(temp_dir)
}
