% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_grab.R
\name{grid_grab}
\alias{grid_grab}
\title{Gridded Data Grabber Grabber - From NTS Grids}
\usage{
grid_grab(aoi_e = NULL, reference_grid, output_directory)
}
\arguments{
\item{aoi_e}{Polygon for the area of interest to be intersected with the NTS grid layer. \emph{(Default: "")} If Default is used the full extent under the reference_grid will be returned.}

\item{reference_grid}{This is a reference raster to provide a projection and a surface to assign values onto, this should be a grid that registers with the other grids you are using for your project. Can be either the location of the raster or a raster object.}

\item{output_directory}{Output directory where your elevation and Wind Ninja elevation grids will be stored.}
}
\description{
This function uses your reference layer to define the NTS grids necessary to pull elevation from the CDEM layers and the national FBP Fuel grid provided by the Government of Canada.
}
\details{
The purpose of this function is to generate a common and rapid elevation layer that is sampled and masked to the reference grid for use within Burn-P3. A second elevation grid is also generated for use in Wind Ninja as that software will fail with NA values in the elevation grid. Fuel data from:
}
\note{
Requires an internet connection
}
\examples{
## Load example data
ref_grid <- rast(system.file("extdata/fuel.tif",package = "BurnP3.HelpR"))
temp_dir <- tempdir()
test <- grid_grab(reference_grid = ref_grid,
                  output_directory = temp_dir)


unlink(temp_dir, recursive = T)
}
\references{
\href{https://open.canada.ca/data/en/dataset/4e66dd2f-5cd0-42fd-b82c-a430044b31de/resource/f645a02d-fbb2-4c35-9bec-8d4ff516f5f8}{National FBP Fuels}

\href{https://open.canada.ca/data/en/dataset/7f245e4d-76c2-4caa-951a-45d1d2051333}{Open Data Canada}

\href{http://ftp.maps.canada.ca/pub/nrcan_rncan/vector/index/}{Open Data Canada FTP}

\href{https://www.firelab.org/project/windninja}{Wind Ninja}
}
