% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_rast.R
\name{hazard_rast}
\alias{hazard_rast}
\title{Wildfire Hazard Raster Generator}
\usage{
hazard_rast(
  stats_file,
  median_pattern,
  mean_pattern,
  burn_count_pattern,
  intensity_directory,
  burn_count_directory
)
}
\arguments{
\item{stats_file}{Character string to the location of the statistics file, if there have been multiple runs aggregated, the aggregate statistics file.}

\item{median_pattern}{Pattern for files that contain median information.}

\item{mean_pattern}{Pattern for files that contain mean information.}

\item{burn_count_pattern}{Pattern for files that contain burn count information.}

\item{intensity_directory}{Directory where intensity rasters can be found}

\item{burn_count_directory}{Directory where burn count rasters can be found}
}
\value{
list
}
\description{
Wildfire Hazard Raster Generator
}
