% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_event_days.R
\name{spread_event_days}
\alias{spread_event_days}
\title{Spread Event Days}
\usage{
spread_event_days(
  input,
  yr_col,
  id_col,
  seasonal = F,
  season_col = "season",
  season_names = "",
  zonal = F,
  zone_col = "wx_zone",
  zone_names = "",
  threshold = 80,
  min_fwi = 19,
  directory = ""
)
}
\arguments{
\item{input}{Weather information file with fire weather index values, the year, and station ids at a minimum.}

\item{yr_col}{Character declaration of year column \emph{(Default = "yr")}}

\item{id_col}{Character declaration of id column \emph{(Default = "id")}}

\item{seasonal}{Declaration of the use of seasons in the weather data set. \emph{(Default = F)}}

\item{season_col}{If seasonal is True, a season column must be declared.}

\item{season_names}{If seasonal is True, identify the season descriptions for use during output.}

\item{zonal}{Declaration of the use of weather zones. \emph{(Default = F)}}

\item{zone_col}{If zonal is True, a zone column must be declared.}

\item{zone_names}{If zonal is True, identify the descriptive names of the zones for use during mapping and output.}

\item{threshold}{For weather based spread event day assessments a threshold is necessary to minimize an excessive tail. This ensures a reasonable distribution for short to mid duration fires but excludes long durations. \emph{(Default = 80)}}

\item{min_fwi}{A minimum fire weather index is used to describe days where fires are more likely to spread and should be consecutively counted. 19 is common in the Canadian Boreal per Podur and Wotton, 2011 \emph{(Default = 19)}}

\item{directory}{Directory for files to be output when using the Burn-P3 directory generator. \emph{(Default = "")}}
}
\value{
data.frame
}
\description{
Spread Event Day calculator based on continuous weather
}
\details{
This is a calculator for spread event days when continuous fire data does not exist and/or fire information from MODIS cannot be used due to inconsistencies. The calculator will use the FWI threshold defined to define the number of consecutive days above that threshold and generate a probability distribution for use in Burn-P3.
}
\examples{

## Load example data
data(weather)

spread_event_days(input = wx_input,
yr_col = "yr",
id_col = "id",
seasonal = FALSE ,
season_col = "season",
season_names = c("Early Spring",
                 "Late Spring",
                 "Early Summer",
                 "Late Summer",
                 "Early Fall",
                 "Late Fall"),
zonal = FALSE,
zone_col = "wx_zone",
zone_names = c("Alpine",
               "Montane",
               "West Alpine",
               "West Montane",
               "West Interior Douglas Fir"),
threshold = 80,
min_fwi = 19,
directory = "")

spread_event_days(input = wx_input,
yr_col = "yr",
id_col = "id",
seasonal = TRUE ,
season_col = "season",
season_names = c("Early Spring",
"Late Spring",
"Early Summer",
"Late Summer",
"Early Fall",
"Late Fall"),
zonal = FALSE,
zone_col = "wx_zone",
zone_names = c("Alpine",
"Montane",
"West Alpine",
"West Montane",
"West Interior Douglas Fir"),
threshold = 80,
min_fwi = 19,
directory = "")

spread_event_days(input = wx_input,
yr_col = "yr",
id_col = "id",
seasonal = FALSE ,
season_col = "season",
season_names = c("Early Spring",
"Late Spring",
"Early Summer",
"Late Summer",
"Early Fall",
"Late Fall"),
zonal = TRUE,
zone_col = "wx_zone",
zone_names = c("Alpine",
"Montane",
"West Alpine",
"West Montane",
"West Interior Douglas Fir"),
threshold = 80,
min_fwi = 19,
directory = "")

}
\references{
\href{cfs.nrcan.gc.ca/publications?id=32563}{Defining fire spread event days for fire-growth modelling. 2011. Podur,J.; Wotton, M. International Journal of Wildland Fire. 20:497-507}
}
